new_project: renamed projects module to project_creation
Also moved the contents of base.py into __init__.py
