+++ s2e_env/commands/project_creation/cgc.py
@@ -20,11 +20,12 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
+
 import datetime
 import os
 
 from s2e_env.command import CommandError
-from .base import BaseProject
+from . import BaseProject
 
 
 class CGCProject(BaseProject):
